Feature extended by Settable
SetValue.java:

Re-Add old SetValue#map for compatibility

Downstream consumers would get NoMethodDefFoundErrors since the newer
version of SetValue#map would have a different signature (first
parameter gets erased to Settable instead of Feature).

Unfortunately, the compiler would fail to resolve calls to
SetValue.map(Feature, Object) now as both functions would be applicable
to those arguments if the Feature happened to be a Settable. The generic
parameters on the deprecated version had to be restricted to not match
anything anyone could pass as an argument to force the compiler to
resolve to the new method when compiling.